requirements.txt
streamlit
pyngrok
llama-cpp-python
huggingface-hub